feat: extract decorator plugin
