+++ blocks/footer/footer.js
@@ -1,11 +1,11 @@
-import { readBlockConfig, decorateIcons } from '../../scripts/lib-franklin.js';
+import { readBlockConfig } from '../../scripts/lib-franklin.js';
 
 /**
  * loads and decorates the footer
  * @param {Element} block The header block element
  */
 
-export default async function decorate(block) {
+export default async function decorate(block, plugins) {
   const cfg = readBlockConfig(block);
   block.textContent = '';
 
@@ -14,6 +14,6 @@ export default async function decorate(block) {
   const html = await resp.text();
   const footer = document.createElement('div');
   footer.innerHTML = html;
-  await decorateIcons(footer);
+  await plugins.decorator.decorateIcons(footer);
   block.append(footer);
 }
