+++ head.html
@@ -1,4 +1,5 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
+<link rel="modulepreload" href="/scripts/plugins/decorator.js"/>
 <link rel="modulepreload" href="/scripts/plugins/rum.js"/>
 <script src="/scripts/lib-franklin.js" type="module" crossorigin="use-credentials"></script>
 <script src="/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
